feat: Final touch ups
